+++ .circleci/config.yml
@@ -34,9 +34,7 @@ workflows:
           matrix: &go-version-matrix
             parameters:
               go_version: ["1.17.9"]
-      - indexer_vs_algod_nightly:
-          name: nightly_test_indexer_vs_algod
-          context: slack-secrets
+
 
 jobs:
   test:
@@ -86,26 +84,6 @@ jobs:
                 }
               ]
             }
-  indexer_vs_algod_nightly:
-    machine:
-      image: << pipeline.parameters.ubuntu_image >>
-    steps:
-      - install_dependencies
-      - run_indexer_vs_algod
-      - slack/notify: &slack-fail-event
-          event: fail
-          custom: |
-            {
-              "blocks": [
-                {
-                  "type": "section",
-                  "text": {
-                    "type": "mrkdwn",
-                    "text": "Failed: <$CIRCLE_BUILD_URL|$CIRCLE_PROJECT_USERNAME/$CIRCLE_PROJECT_REPONAME>\n(<https://app.circleci.com/pipelines/github/$CIRCLE_PROJECT_USERNAME/$CIRCLE_PROJECT_REPONAME?branch=$CIRCLE_BRANCH|$CIRCLE_BRANCH>) \n- <$CIRCLE_BUILD_URL|$CIRCLE_JOB> failed\n\nThis is probably a result of `make indexer-v-algod`\nSee <https://algorand.atlassian.net/wiki/spaces/LAMPREY/pages/2339536905/Nightly+Indexer+Tests#Q%3A-What-does-it-mean-that-the-nightly-test-failed-because-of-make-indexer-v-algod%3F|this wiki page> for more details"
-                  }
-                }
-              ]
-            }
 
 commands:
   install_dependencies:
@@ -120,12 +98,6 @@ commands:
             sudo apt -y install python3 python3-pip python3-setuptools python3-wheel libboost-math-dev libffi-dev
             pip3 install -r misc/requirements.txt
             pip3 install e2e_tests/
-
-      - run:
-          name: sync submodules (go-algorand)
-          command: |
-            git submodule sync
-            git submodule update --init
 
       - run: echo 'export PATH=$PATH:/usr/local/go/bin:$HOME/.local/bin' >> $BASH_ENV
 
@@ -173,6 +145,3 @@ commands:
       - run: make test-generate
       - run: make fakepackage
 
-  run_indexer_vs_algod:
-    steps:
-      - run: make indexer-v-algod
